fix : 이미지 갖는 객체 Course -> Session 으로 이동
