+++ src/main/java/nextstep/courses/domain/Course.java
@@ -13,20 +13,12 @@ public class Course {
 
     private Sessions sessions;
 
-    private SessionImage sessionImage;
-
     private Long creatorId;
 
     private LocalDateTime createdAt;
 
     private LocalDateTime updatedAt;
 
-
-    public Course imageOf(SessionImage sessionImage) {
-        Course course = this;
-        course.sessionImage = sessionImage;
-        return course;
-    }
     public Course() {
     }
 
@@ -37,7 +29,6 @@ public class Course {
     public Course(Long id, String title, Long creatorId, LocalDateTime createdAt, LocalDateTime updatedAt) {
         this.id = id;
         this.title = title;
-        this.sessionImage = null;
         this.groups = Collections.emptyList();
         this.sessions = Sessions.initialize();
         this.creatorId = creatorId;
@@ -48,10 +39,6 @@ public class Course {
     public void addSession(Session session) {
         session.toCourse(this);
         this.sessions = sessions.addSession(session);
-    }
-
-    public boolean hasImage() {
-        return !(sessionImage == null);
     }
 
     public String getTitle() {
