+++ src/test/java/nextstep/courses/domain/CourseTest.java
@@ -26,15 +26,4 @@ public class CourseTest {
         course.addSession(session2);
         assertThat(course.getSessions().size()).isEqualTo(2);
     }
-
-    @Test
-    @DisplayName("강의는 강의 커버 이미지 정보를 가진다.")
-    void saveSessionImageTest() throws InvalidImageFormatException {
-        assertThat(course.hasImage()).isFalse();
-
-        SessionImage 강의_이미지 = SessionImage.nameOf("강의 이미지");
-        course.imageOf(강의_이미지);
-
-        assertThat(course.hasImage()).isTrue();
-    }
 }
